Fix scrolling of long modal dialogs
NOTE: react-modal removed the ability to select where the modal div is inserted into in version 1.3.0. This is because aria-hidden (i.e. "ignore this" for screenreaders) is added to the rest of the app, but now that the modal is also included in the app, it gets hidden from screenreaders as well. There's currently a PR open (https://github.com/reactjs/react-modal/pull/183) that tries to address this, so we might see the return of this feature at some point. We should upgrade then.